+++ Packages/NativeGamepad/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.20.10] - 2026-06-02
+
+- No functional changes (version synchronized with the monorepo release).
+
 ## [0.20.9] - 2026-06-02
 
 - No functional changes (version synchronized with the monorepo release).